Add user parameter read/write
A page in high memory is used to store 32 bytes of "user parameters".  These would be things like autonomous modes or PID constants.  These parameters are non-volatile across power cycles and user program downloads but will be cleared when the ROBOTC master firmware is updated.

+++ stm32_flash.c
@@ -317,7 +317,7 @@ FLASH_Status FLASH_ProgramWord(uint32_t Address, uint32_t Data)
       half word */
       Addr++;
 
-      *Addr = Data >> 16;
+      *Addr = (uint16_t)(Data >> 16);
 
       /* Wait for last operation to be completed */
       status = FLASH_WaitForLastOperation(ProgramTimeout);
